[REFACTOR] 날짜계산 함수를 utils로 빼서 정리
